% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_disputes}
\alias{age_disputes}
\title{Age disputes}
\format{
A data frame with 3,354 rows and 8 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality of people applying}
  \item{Region}{World region the nation is in}
  \item{Raised or resolved}{'Raised' or 'Resolved'}
  \item{Raised type / Resolved outcome}{'Existing asylum application', 'Asylum application raised in quarter', 'Less than 18', or '18+'}
  \item{Age disputes}{Number of age disputes}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/asylum-and-resettlement-datasets}
}
\usage{
age_disputes
}
\description{
Age disputes raised and outcomes of age disputes.
}
\keyword{datasets}
