% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_of_asylum_operations}
\alias{age_of_asylum_operations}
\title{Age of asylum operations}
\format{
A data frame with 23 rows and 8 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Total}{Total awaiting initial asylum decision}
  \item{Less than 3 months}{Number waiting less than three months}
  \item{3-6 months}{Number waiting 3-6 months}
  \item{6-12 months}{Number waiting 6-12 months}
  \item{12 months+}{Number waiting more than a year}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
age_of_asylum_operations
}
\description{
Age of Asylum Operations Initial Decision Work In Progress - applications
made from 1st Oct 2006.
}
\keyword{datasets}
