% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{appeal_representation_rate}
\alias{appeal_representation_rate}
\title{Appeal Representation Rates}
\format{
A data frame with 42 rows and 7 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{All hearings (\%)}{All hearings}
  \item{First Tier (\%)}{First tier}
  \item{Upper Tier (\%)}{Upper tier}
  \item{Deportation (\%)}{Deportation}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
appeal_representation_rate
}
\description{
Appeal representation rates.
}
\keyword{datasets}
