% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{appeals_determined}
\alias{appeals_determined}
\title{Asylum appeals determined}
\format{
A data frame with 11,973 rows and 8 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality of people applying}
  \item{Region}{World region the nation is in}
  \item{Outcome}{'Allowed', 'Dismissed', or 'Withdrawn'}
  \item{Sex}{Gender/sex of applicants}
  \item{Appeals determined}{Number of appeals determined}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/asylum-and-resettlement-datasets}
}
\usage{
appeals_determined
}
\description{
Outcomes of asylum appeals raised at the First-Tier Tribunal, by nationality
and sex.
}
\keyword{datasets}
