% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{appeals_lodged}
\alias{appeals_lodged}
\title{Asylum appeals lodged}
\format{
A data frame with 6,621 rows and 7 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality of people applying}
  \item{Region}{World region the nation is in}
  \item{Sex}{Gender/sex of applicants}
  \item{Appeals lodged}{Number of appeals lodged}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/asylum-and-resettlement-datasets}
}
\usage{
appeals_lodged
}
\description{
Asylum appeals raised at the First-Tier Tribunal, by nationality and sex.
}
\keyword{datasets}
