% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{applications}
\alias{applications}
\title{Asylum applications}
\format{
A data frame with 57,482 rows and 11 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality of people applying}
  \item{Region}{World region the nation is in}
  \item{Age}{Age group of applicants}
  \item{Sex}{Gender/sex of applicants}
  \item{Applicant type}{'Main applicant' or 'Dependant'}
  \item{UASC}{Are they an unaccompanied asylum-seeking child?}
  \item{Location of application}{'At Port' or 'In Country'}
  \item{Applications}{Number of applications}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/asylum-and-resettlement-datasets}
}
\usage{
applications
}
\description{
Asylum applications raised, by nationality, age, sex, UASC, applicant type,
and location of application.
}
\keyword{datasets}
