% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{applications_processed_in_6_months}
\alias{applications_processed_in_6_months}
\title{Asylum applications processed within 6 months}
\format{
A data frame with 33 rows and 16 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Decisions in}{Column just containing a label saying "6 months"}
  \item{Total Applications Received}{Total applications received}
  \item{Of those Applications received the number completed within six months}{Number of applications completed}
  \item{Of those Applications received the percentage completed within six months}{Percent of applications completed}
  \item{Total Male Applications Received}{Total applications received for males}
  \item{Of those Male Applications received the number completed within six months}{Number of applications completed for males}
  \item{Of those Male Applications received the percentage completed within six months}{Percent of applications completed for males}
  \item{Total Female Applications Received}{Total applications received for females}
  \item{Of those Female Applications received the number completed within six months}{Number of applications completed for females}
  \item{Of those Female Applications received the percentage completed within six months}{Percent of applications completed for females}
  \item{Total Unknown Applications Received}{Total applications received for people whose sex/gender is unknown}
  \item{Of those Unknown Applications received the number completed within six months}{Number of applications completed for people whose gender/sex is unknown}
  \item{Of those Unknown Straightforward Applications received the percentage completed within six months}{Percent of applications completed for people whose gender/sex is unknown}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
applications_processed_in_6_months
}
\description{
Percentage of Asylum applications processed within 6 months
}
\keyword{datasets}
