% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aspen}
\alias{aspen}
\title{Aspen Cards in use}
\format{
A data frame with 17 rows and 4 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Aspen Cards in use}{Number of Aspen Cards in use}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
aspen
}
\description{
The number of Aspen Cards in use.
}
\keyword{datasets}
