% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aspen_support}
\alias{aspen_support}
\title{Aspen Card-related support}
\format{
A data frame with 17 rows and 6 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Home Office}{NUmber of support enquiries received by the Home Office}
  \item{Sodexo}{enquiries received by Sodexo}
  \item{Migrant Help}{enquiries received by Migrant Help}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
aspen_support
}
\description{
The number of Aspen Card-related support enquiries received by the Home
Office, Sodexo and Migrant Help.
}
\keyword{datasets}
