% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asylum-package.R
\docType{package}
\name{asylum-package}
\alias{asylum}
\alias{asylum-package}
\title{asylum: Data on Asylum and Resettlement for the UK}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Data on Asylum and Resettlement for the UK, provided by the Home Office \url{https://www.gov.uk/government/statistical-data-sets/immigration-system-statistics-data-tables}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://humaniverse.github.io/asylum/}
  \item \url{https://github.com/humaniverse/asylum}
  \item Report bugs at \url{https://github.com/humaniverse/asylum/issues}
}

}
\author{
\strong{Maintainer}: Matthew Gwynfryn Thomas \email{matthewgthomas@gmail.com}

Authors:
\itemize{
  \item Mike Page \email{mpage@redcross.org.uk}
  \item Jenna Tan \email{JennaTan@redcross.org.uk}
}

Other contributors:
\itemize{
  \item  British Red Cross [copyright holder]
}

}
\keyword{internal}
