% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asylum_costs_and_productivity}
\alias{asylum_costs_and_productivity}
\title{Breakdown of Asylum costs and productivity}
\format{
A data frame with 12 rows and 12 variables:
\describe{
  \item{Financial Year}{Financial year}
  \item{Total Asylum Costs}{The costs of managing an asylum operations claim include a) costs associated with deciding a case (screening clients, interviews and issuing a decision), b) managing any related appeal, c) the cost of all asylum operations support (S4, S95 & S98) d) detention costs where detention has been used and e) enforcement costs (escorting and assisted voluntary returns). Also included are the costs of department staff and buildings and the proportionate costs of the Home Office building and managerial costs}
  \item{Asylum WIP (new method)}{Number of work in progress claims (new method)}
  \item{Asylum WIP (old method)}{Number of work in progress claims (old method)}
  \item{Unit Cost}{Unit Costs - This is based on the Work in Progress figure. The costs of an asylum operations case will vary considerably dependent on the type of case (single, dependents) as well as barriers to documentation, legal challenges and logistical and practical challenges}
  \item{Total Conclusion}{Total conclusions}
  \item{Average Conclusions Per Month}{Average conclusions per month}
  \item{Conclusions Productivity}{Conclusions producitivity}
  \item{Total Principal Stages Completed}{Total principal stages completed}
  \item{Average Principal Stages Completed Per Month}{Average principal stages completed per month}
  \item{Asylum Caseworking Staff}{Number of asylum caseworking staff}
  \item{Productivity}{Productivity of staff}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
asylum_costs_and_productivity
}
\description{
Breakdown of Asylum costs and productivity.
}
\keyword{datasets}
