% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asylum_productivity_breakdown}
\alias{asylum_productivity_breakdown}
\title{Breakdown of asylum productivity}
\format{
A data frame with 6 variables:
\describe{
  \item{Date}{Date}
  \item{Initial Decisions}{Number of initial decisions made}
  \item{Substantive Interviews}{Number of substantive interviews completed}
  \item{Total Principal Stages Completed}{Total number of initial decisions and substantive interviews}
  \item{Asylum Caseworking Staff}{Number of asylum caseworking staff}
  \item{Productivity}{Productivity of staff (total principal stages completed divided by asylum caseworking staff)}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
asylum_productivity_breakdown
}
\description{
Breakdown of asylum productivity by month.
}
\keyword{datasets}
