% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asylum_work_in_progress}
\alias{asylum_work_in_progress}
\title{Asylum work in progress}
\format{
A data frame with 12 rows and 17 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Total Work In Progress}{Number of cases that are not concluded at the given point in time}
  \item{Awaiting Initial Asylum Decision}{Number waiting an initial decision}
  \item{Post Decision}{The 'Post Decision' sub-category includes cases that have had a decision but have not appealed, are not appeal rights exhausted, and have not been removed}
  \item{Asylum Appeal Outstanding}{Number with appeals outstanding}
  \item{Subject to Removal Action}{Number of cases in the work in progress caseload have been processed through the initial stages of the asylum operations system and are now 'subject to removal action'}
  \item{On Hold}{'On Hold' sub-category includes pre- and post-decision absconders}
  \item{Further Leave Application Outstanding}{Further Leave applications from former Unaccompanied Asylum Seeking Children given Discretionary Leave}
  \item{Case Age: 0:12 Months}{Age of the case}
  \item{Case Age: 12:24 Months}{Age of the case}
  \item{Case Age: 24:36 Months}{Age of the case}
  \item{Case Age: 36+ Months}{Age of the case}
  \item{Gender: Male}{Number of males}
  \item{Gender: Female}{Number of females}
  \item{Gender: Unknown}{Number of people whose gender/sex is unknown}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
asylum_work_in_progress
}
\description{
Asylum work in progress.
}
\keyword{datasets}
