% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{children_entering_detention}
\alias{children_entering_detention}
\title{Children entering detention}
\format{
A data frame with 13 rows and 7 variables:
\describe{
  \item{Date of entry to detention}{Date of entry to detention}
  \item{Total children}{Total children}
  \item{Children 4 and under}{Children 4 and under}
  \item{Children 5 to 11 years}{Children 5 to 11 years}
  \item{Children 12 to 16 years}{Children 12 to 16 years}
  \item{Children 17 years}{Children 17 years}
  \item{Total adults 18 and over}{Total adults (18 and over)}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/immigration-system-statistics-data-tables-year-ending-march-2023#detention}
}
\usage{
children_entering_detention
}
\description{
Number of occurrences of children entering detention, 2010 to 2022
}
\keyword{datasets}
