% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{claims_certified_section_94}
\alias{claims_certified_section_94}
\title{Asylum claims certified under Section 94}
\format{
A data frame with 2,142 rows and 10 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality}
  \item{Region}{Region}
  \item{Initial decisions from designated states}{Initial decisions from designated states}
  \item{Refusals from designated states}{Refusals from designated states}
  \item{Clearly unfounded refusals (designated states)}{Clearly unfounded refusals (designated states)}
  \item{Clearly unfounded refusals (non-designated states)}{Clearly unfounded refusals (non-designated states)}
  \item{Total clearly unfounded refusals}{Total clearly unfounded refusals}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/asylum-and-resettlement-datasets}
}
\usage{
claims_certified_section_94
}
\description{
Initial decisions on asylum applications certified under Section 94, by
nationality
}
\keyword{datasets}
