% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decision_quality}
\alias{decision_quality}
\title{Decision Quality}
\format{
A data frame with 15 rows and 5 variables:
\describe{
  \item{Financial Year}{Financial year}
  \item{Assurance}{First line or second line assurance}
  \item{Decision Quality (\%)}{Decision Quality percentage}
  \item{Number of Decisions Sampled}{Number of decisions sampled}
  \item{Sample Size (\%)}{Sample size}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
decision_quality
}
\description{
The Decision Quality percentage of Decisions Sampled.
}
\keyword{datasets}
