% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decisions_resettlement}
\alias{decisions_resettlement}
\title{Initial decisions and resettlement}
\format{
A data frame with 95,179 rows and 14 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality of people applying}
  \item{Region}{World region the nation is in}
  \item{Case type}{'Asylum Case' or 'Resettlement Case'}
  \item{Case outcome group}{'Grant of Other Leave', 'Grant of Protection', 'Refused', or 'Withdrawn'}
  \item{Case outcome}{Detailed description of the case outcome}
  \item{Age}{Age group of applicants}
  \item{Sex}{Gender/sex of applicants}
  \item{Applicant type}{'Main applicant' or 'Dependant'}
  \item{UASC}{Are they an unaccompanied asylum-seeking child?}
  \item{Host Country}{Host country, in cases of resettlement}
  \item{Decisions}{Number of decisions}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/asylum-and-resettlement-datasets}
}
\usage{
decisions_resettlement
}
\description{
Outcomes of asylum applications at initial decision, and refugees resettled
in the UK, by nationality, age, sex, applicant type, and UASC (unaccompanied
asylum-seeking children).
}
\keyword{datasets}
