% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detention_cost_per_day}
\alias{detention_cost_per_day}
\title{Average cost per day to hold an individual in immigration detention}
\format{
A data frame with 4 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Cost}{Cost}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#immigration-enforcement}
}
\usage{
detention_cost_per_day
}
\description{
Average cost per day to hold an individual in immigration detention
}
\keyword{datasets}
