% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detention_pregnant_women}
\alias{detention_pregnant_women}
\title{Number of pregnant women detained in the immigration detention estate}
\format{
A data frame with 29 rows and 4 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Number of pregnant women detained in the immigration detention estate}{Number of pregnant women detained in the immigration detention estate}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#immigration-enforcement}
}
\usage{
detention_pregnant_women
}
\description{
Number of pregnant women detained in the immigration detention estate
}
\keyword{datasets}
