% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detention_summary_flow}
\alias{detention_summary_flow}
\title{People entering, leaving and in detention by asylum and non-asylum}
\format{
A data frame with 13 rows and 10 variables:
\describe{
  \item{Date}{Date}
  \item{Entering detention}{Entering detention}
  \item{Entering detention Asylum}{Entering detention Asylum}
  \item{Entering detention Non-asylum}{Entering detention Non-asylum}
  \item{Leaving detention}{Leaving detention}
  \item{Leaving detention Asylum}{Leaving detention Asylum}
  \item{Leaving detention Non-asylum}{Leaving detention Non-asylum}
  \item{In detention (at end of period)}{In detention (at end of period)}
  \item{In detention Asylum}{In detention Asylum}
  \item{In detentionNon-asylum}{In detention Non-asylum}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/immigration-system-statistics-data-tables-year-ending-december-2022#detention}
}
\usage{
detention_summary_flow
}
\description{
People entering, leaving and in detention by asylum and non-asylum, 2010 to 2022.
}
\keyword{datasets}
