% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detentions_deaths_and_absconds}
\alias{detentions_deaths_and_absconds}
\title{Deaths of persons detained under Immigration Act Powers and absconds from
detention, 2017 to 2022}
\format{
A data frame with 6 rows and 5 variables:
\describe{
  \item{Year}{Year}
  \item{Deaths of persons detained under Immigration Powers}{Deaths of persons detained under Immigration Powers}
  \item{Of which occurred in the detention estate}{Of which occurred in the detention estate}
  \item{Escapes from detention centres}{Escapes from detention centres}
  \item{Absconds from escort}{Absconds from escort}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/immigration-system-statistics-data-tables-year-ending-december-2022#detention}
}
\usage{
detentions_deaths_and_absconds
}
\description{
Deaths of persons detained under Immigration Act Powers and absconds from
detention, 2017 to 2022
}
\keyword{datasets}
