% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detentions_deaths_details}
\alias{detentions_deaths_details}
\title{Deaths of persons detained under Immigration Act Powers, 2017 to 2021}
\format{
A data frame with 10 rows and 8 variables:
\describe{
  \item{Year}{Year}
  \item{Number}{Number identifying a person within each year}
  \item{Sex}{Sex}
  \item{Age}{Age}
  \item{Nationality}{Nationality}
  \item{Cause of death}{Cause of death}
  \item{Place of incident}{Place of incident}
  \item{Place of death}{Place of death}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/immigration-system-statistics-data-tables-year-ending-december-2022#detention}
}
\usage{
detentions_deaths_details
}
\description{
Deaths of persons detained under Immigration Act Powers, 2017 to 2021
}
\keyword{datasets}
