% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stats.R
\name{download_stats}
\alias{download_stats}
\title{Try to download a statistics data file from the Home Office website.}
\usage{
download_stats(url, name_of_file)
}
\arguments{
\item{url}{A URL with a set of links to datasets}

\item{name_of_file}{Text of the link to the file}
}
\value{
A character string containing the location of the downloaded file
}
\description{
Try to download a statistics data file from the Home Office website.
}
