% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dublin_regulation}
\alias{dublin_regulation}
\title{Dublin regulation}
\format{
A data frame with 3,508 rows and 8 variables:
\describe{
  \item{Date}{The period in which the request or transfer of an asylum applicant under the Dublin Regulation was made}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Member State}{As applicable, the EU Member State which either raises, or receives, a request or transfer of an asylum applicant under the Dublin Regulation}
  \item{In / Out}{The direction of a request or transfer: either into, or out of, the UK}
  \item{Request / Transfer}{Whether the case is a request for transfer, or a transfer, of an asylum applicant}
  \item{Article}{The Article (rule) of the Dublin Regulation under which the request or transfer is carried out}
  \item{Count}{The number of transfers, or requests to transfer under the Dublin Regulation}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/asylum-and-resettlement-datasets}
}
\usage{
dublin_regulation
}
\description{
Transfers and requests for transfer under the Dublin regulation, by EU
member state and article.
}
\keyword{datasets}
