% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{family_reunion}
\alias{family_reunion}
\title{Family reunion visa grants}
\format{
A data frame with 7,475 rows and 8 variables:
\describe{
  \item{Date}{The period in which the Family Reunion visa was granted}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality of the visa recipient}
  \item{Region}{World region of the country of nationality for the visa recipient}
  \item{Sex}{The sex of the individual}
  \item{Age}{Age of individual at time of visa grant}
  \item{Visas granted}{The number of family reunion visas granted}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/asylum-and-resettlement-datasets}
}
\usage{
family_reunion
}
\description{
Family reunion visas granted to family members of refugees, by nationality,
sex and age.
}
\keyword{datasets}
