% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{fetch_applications}
\alias{fetch_applications}
\title{Load latest data on asylum applications}
\usage{
fetch_applications()
}
\value{
A tibble containing the wrangled data
}
\description{
Load latest data on asylum applications
}
