% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{fetch_asylum_support}
\alias{fetch_asylum_support}
\title{Load latest data on people receiving asylum support by Local Authority and Region}
\usage{
fetch_asylum_support()
}
\value{
A tibble containing the wrangled data
}
\description{
Load latest data on people receiving asylum support by Local Authority and Region
}
