% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{fetch_awaiting_decision}
\alias{fetch_awaiting_decision}
\title{Load latest data on people awaiting a decision on their asylum claim}
\usage{
fetch_awaiting_decision()
}
\value{
A tibble containing the wrangled data
}
\description{
Load latest data on people awaiting a decision on their asylum claim
}
