% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{fetch_decisions}
\alias{fetch_decisions}
\title{Load latest data on initial decisions}
\usage{
fetch_decisions()
}
\value{
A tibble containing the wrangled data
}
\description{
Load latest data on initial decisions
}
