% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{fetch_irregular_migration}
\alias{fetch_irregular_migration}
\title{Load latest data on irregular migration}
\usage{
fetch_irregular_migration()
}
\value{
A tibble containing the wrangled data
}
\description{
Load latest data on irregular migration
}
