% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{fetch_local_authority_support}
\alias{fetch_local_authority_support}
\title{Load latest data on people receiving asylum support by local authority}
\usage{
fetch_local_authority_support()
}
\value{
A tibble containing the wrangled data
}
\description{
Load latest data on people receiving asylum support by local authority
}
