% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{fetch_reunion}
\alias{fetch_reunion}
\title{Load latest data on family reunion visa grants}
\usage{
fetch_reunion()
}
\value{
A tibble containing the wrangled data
}
\description{
Load latest data on family reunion visa grants
}
