% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{fetch_small_boat_asylum_claims}
\alias{fetch_small_boat_asylum_claims}
\title{Load latest data on asylum applications from small boat arrivals}
\usage{
fetch_small_boat_asylum_claims()
}
\value{
A tibble containing the wrangled data
}
\description{
Load latest data on asylum applications from small boat arrivals
}
