% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grant_rates_initial_annual}
\alias{grant_rates_initial_annual}
\title{Grant rates at initial decision (annual)}
\format{
A data frame with 3,295 rows and 6 variables:
\describe{
  \item{Year}{Year}
  \item{Nationality}{Nationality}
  \item{Region}{Region}
  \item{Grant}{Total number of grants at initial decision}
  \item{Refused}{Total number of refusals at initial decision}
  \item{Initial grant rate}{Initial grant rate - percentage}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/immigration-system-statistics-data-tables-year-ending-december-2022}
}
\usage{
grant_rates_initial_annual
}
\description{
Grant rate is the proportion of initial decisions which resulted in a grant
of protection or other leave.
}
\keyword{datasets}
