% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irregular_migration}
\alias{irregular_migration}
\title{Irregular migration to the UK}
\format{
A data frame with 6,503 rows and 9 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Method of entry}{Method of entry}
  \item{Nationality}{Nationality}
  \item{Region}{Region}
  \item{Sex}{Sex}
  \item{Age Group}{Age Group}
  \item{Number of detections}{Number of detections}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/immigration-system-statistics-data-tables-year-ending-december-2022#detention}
}
\usage{
irregular_migration
}
\description{
Number of detected attempts to enter the UK irregularly by method of entry,
nationality, region, sex and age group.
}
\keyword{datasets}
