% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{local_authority_resettlement}
\alias{local_authority_resettlement}
\title{Resettlement by local authority}
\format{
A data frame with 2,464 rows and 8 variables:
\describe{
  \item{Date}{The period in which the refugee arrived in the UK}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{UK Region}{Region of the UK in which the refugee was initially resettled}
  \item{LAD Code}{Local Authority (district) code as supplied by ONS}
  \item{Local Authority}{The Local Authority in which the refugee was initially resettled}
  \item{Resettlement Scheme}{The scheme under which the refugee was resettled}
  \item{Persons}{The number of refugees resettled}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/asylum-and-resettlement-datasets}
}
\usage{
local_authority_resettlement
}
\description{
Refugees resettled by resettlement scheme and local authority.
}
\keyword{datasets}
