% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{notices_of_intent}
\alias{notices_of_intent}
\title{Notices of intent issued, by nationality}
\format{
A data frame with 11 rows and 2 variables:
\describe{
  \item{Nationality}{Nationality}
  \item{Total}{Total number of notices of intent issued}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/immigration-system-statistics-data-tables-year-ending-december-2022}
}
\usage{
notices_of_intent
}
\description{
Notices of intent issued, by nationality, 1 January 2021 to 30 September 2022
}
\details{
1. Data on inadmissibility are taken from internal Home Office management
information and should be considered provisional.
2. Includes main applicants and dependants.
3. Data for October to December 2022 was not available at the time of data
extraction for this release, due to ongoing work as immigration  data
transitions to a new case working system. Updated data will be included in a
future edition of the Immigration System Statistics release.
}
\keyword{datasets}
