% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrm_conclusive_grounds}
\alias{nrm_conclusive_grounds}
\title{NRM conclusive grounds decisions}
\format{
A data frame with 47 rows and 12 variables:
\describe{
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Adult (18 or over) - Negative conclusive grounds}{Adult (18 or over) - Negative conclusive grounds}
  \item{Adult (18 or over) - Positive conclusive grounds}{Adult (18 or over) - Positive conclusive grounds}
  \item{Adult (18 or over) - Total}{Adult (18 or over) - Total}
  \item{Child (17 or under) - Negative conclusive grounds}{Child (17 or under) - Negative conclusive grounds}
  \item{Child (17 or under) - Positive conclusive grounds}{Child (17 or under) - Positive conclusive grounds}
  \item{Child (17 or under) - Total}{Child (17 or under) - Total}
  \item{Age not specified or unknown - Negative conclusive grounds}{Age not specified or unknown - Negative conclusive grounds}
  \item{Age not specified or unknown - Positive conclusive grounds}{Age not specified or unknown - Positive conclusive grounds}
  \item{Age not specified or unknown - Total}{Age not specified or unknown - Total}
  \item{Total}{Total decisions}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/national-referral-mechanism-statistics}
}
\usage{
nrm_conclusive_grounds
}
\description{
NRM conclusive grounds decisions made by both competent authorities by
quarter, outcome and age at exploitation.
}
\keyword{datasets}
