% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrm_duty_to_notify_2022_q3}
\alias{nrm_duty_to_notify_2022_q3}
\title{Referrals received via the Duty to Notify process in Q3 2022, by nationality}
\format{
A data frame with 2 variables:
\describe{
  \item{Nationality}{Nationality}
  \item{Total}{Number of referrals via duty to notify}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/national-referral-mechanism-statistics}
}
\usage{
nrm_duty_to_notify_2022_q3
}
\description{
Referrals received via the Duty to Notify process in Q3 2022, by nationality
}
\keyword{datasets}
