% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrm_duty_to_notify_longitudinal}
\alias{nrm_duty_to_notify_longitudinal}
\title{Referrals received via the Duty to Notify process, by quarter}
\format{
A data frame with 41 rows and 3 variables:
\describe{
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Total}{Number of referrals via duty to notify}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/national-referral-mechanism-statistics}
}
\usage{
nrm_duty_to_notify_longitudinal
}
\description{
Referrals received via the Duty to Notify process, by quarter
}
\keyword{datasets}
