% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrm_reasonable_grounds}
\alias{nrm_reasonable_grounds}
\title{NRM reasonable grounds decisions}
\format{
A data frame with 12 variables:
\describe{
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Adult (18 or over) - Negative reasonable grounds}{Adult (18 or over) - Negative reasonable grounds}
  \item{Adult (18 or over) - Positive reasonable grounds}{Adult (18 or over) - Positive reasonable grounds}
  \item{Adult (18 or over) - Total}{Adult (18 or over) - Total}
  \item{Child (17 or under) - Negative reasonable grounds}{Child (17 or under) - Negative reasonable grounds}
  \item{Child (17 or under) - Positive reasonable grounds}{Child (17 or under) - Positive reasonable grounds}
  \item{Child (17 or under) - Total}{Child (17 or under) - Total}
  \item{Age not specified or unknown - Negative reasonable grounds}{Age not specified or unknown - Negative reasonable grounds}
  \item{Age not specified or unknown - Positive reasonable grounds}{Age not specified or unknown - Positive reasonable grounds}
  \item{Age not specified or unknown - Total}{Age not specified or unknown - Total}
  \item{Total}{Total decisions}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/national-referral-mechanism-statistics}
}
\usage{
nrm_reasonable_grounds
}
\description{
NRM reasonable grounds decisions made by both competent authorities by
quarter, outcome and age at exploitation.
}
\keyword{datasets}
