% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrm_referrals_2023_q2}
\alias{nrm_referrals_2023_q2}
\title{NRM referrals - Q2 2023}
\format{
A data frame with 49,752 rows and 7 variables:
\describe{
  \item{First responder type}{First responder type}
  \item{First responder}{First responder}
  \item{Exploitation type}{Labour, domestic, sexual etc. exploitation}
  \item{Age at exploitation}{Age group}
  \item{Gender}{Gender/sex}
  \item{Nationality}{Nationality}
  \item{People}{Number of people referred to NRM}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/national-referral-mechanism-statistics}
}
\usage{
nrm_referrals_2023_q2
}
\description{
NRM referrals - Q2 2023
}
\keyword{datasets}
