% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrm_referrals_longitudinal}
\alias{nrm_referrals_longitudinal}
\title{NRM referrals by quarter by age at exploitation and location of exploitation}
\format{
A data frame with 870 rows and 5 variables:
\describe{
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Age at exploitation}{Age at exploitation}
  \item{Location of exploitation}{Location of exploitation}
  \item{NRM referrals}{Number of NRM referrals}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/national-referral-mechanism-statistics}
}
\usage{
nrm_referrals_longitudinal
}
\description{
NRM referrals by quarter by age at exploitation and location of exploitation
}
\keyword{datasets}
