% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrpf_change_of_conditions_age}
\alias{nrpf_change_of_conditions_age}
\title{NRPF - Destitution Change of Conditions Application by Age Group}
\format{
A data frame with 20 rows and 12 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Under 18}{Number of applicants under 18}
  \item{18 to 25}{Number of applicants aged 18-25}
  \item{26 to 30}{Number of applicants aged 26-30}
  \item{31 to 40}{Number of applicants aged 31-40}
  \item{41 to 50}{Number of applicants aged 41-50}
  \item{51 to 60}{Number of applicants aged 51-60}
  \item{61 to 70}{Number of applicants aged 61-70}
  \item{71 to 80}{Number of applicants aged 71-80}
  \item{Over 80}{Number of applicants aged over 80}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
nrpf_change_of_conditions_age
}
\description{
NFRP - Destitution Change of Conditions Applications by Age Group. Data
relating to people who are on a Family or Human Rights route and apply to
have the No Recourse to Public Funds (NRPF) condition lifted. A factsheet
relating to NRPF can be found here:
https://homeofficemedia.blog.gov.uk/2020/05/05/no-recourse-to-public-funds-nrpf/
}
\keyword{datasets}
