% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrpf_change_of_conditions_decisions}
\alias{nrpf_change_of_conditions_decisions}
\title{NFRP - Destitution Change of Conditions Applications and Outcomes}
\format{
A data frame with 20 rows and 10 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Applications Received}{Number of applications received}
  \item{Of which: Pending}{Number of applications pending}
  \item{Decisions}{Number of decisions made}
  \item{Of which: Accepted}{Decisions that were accepted}
  \item{Of which: Rejected}{Decisions that were rejected}
  \item{Average Days to Decision}{Average days taken to reach a decision}
  \item{Acceptance Rate}{Rate of acceptance}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
nrpf_change_of_conditions_decisions
}
\description{
NFRP - Destitution Change of Conditions Applications and Outcomes. Data
relating to people who are on a Family or Human Rights route and apply to
have the No Recourse to Public Funds (NRPF) condition lifted. A factsheet
relating to NRPF can be found here:
https://homeofficemedia.blog.gov.uk/2020/05/05/no-recourse-to-public-funds-nrpf/
}
\keyword{datasets}
