% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrpf_change_of_conditions_gender}
\alias{nrpf_change_of_conditions_gender}
\title{NFRP - Destitution Change of Conditions Applications by Gender}
\format{
A data frame with 20 rows and 6 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Male}{Number of applications from males}
  \item{Female}{Number of applications from females}
  \item{Recorded as Unknown}{Number of applications from people whose sex/gender is unknown}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
nrpf_change_of_conditions_gender
}
\description{
NFRP - Destitution Change of Conditions Applications by Gender. Data
relating to people who are on a Family or Human Rights route and apply to
have the No Recourse to Public Funds (NRPF) condition lifted. A factsheet
relating to NRPF can be found here:
https://homeofficemedia.blog.gov.uk/2020/05/05/no-recourse-to-public-funds-nrpf/
}
\keyword{datasets}
