% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrpf_change_of_conditions_nationality}
\alias{nrpf_change_of_conditions_nationality}
\title{NRPF - Destitution Change of Conditions Application by Nationality}
\format{
A data frame with 154 rows and 21 variables:
\describe{
  \item{Nationality}{Nationality}
  \item{2017 Q3}{2017 Q3}
  \item{2017 Q4}{2017 Q4}
  \item{2018 Q1}{2018 Q1}
  \item{2018 Q2}{2018 Q2}
  \item{2018 Q3}{2018 Q3}
  \item{2018 Q4}{2018 Q4}
  \item{2019 Q1}{2019 Q1}
  \item{2019 Q2}{2019 Q2}
  \item{2019 Q3}{2019 Q3}
  \item{2019 Q4}{2019 Q4}
  \item{2020 Q1}{2020 Q1}
  \item{2020 Q2}{2020 Q2}
  \item{2020 Q3}{2020 Q3}
  \item{2020 Q4}{2020 Q4}
  \item{2021 Q1}{2021 Q1}
  \item{2021 Q2}{2021 Q2}
  \item{2021 Q3}{2021 Q3}
  \item{2021 Q4}{2021 Q4}
  \item{2022 Q1}{2022 Q1}
  \item{2022 Q2}{2022 Q2}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
nrpf_change_of_conditions_nationality
}
\description{
NFRP - Destitution Change of Conditions Applications by Nationality. Data
relating to people who are on a Family or Human Rights route and apply to
have the No Recourse to Public Funds (NRPF) condition lifted. A factsheet
relating to NRPF can be found here:
https://homeofficemedia.blog.gov.uk/2020/05/05/no-recourse-to-public-funds-nrpf/
}
\keyword{datasets}
