% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nts}
\alias{nts}
\title{National Transfer Scheme}
\format{
A data frame with 10,393 rows and 9 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Region}{UK region}
  \item{Local Authority or HSCT (NI)}{Local Authority, or Health & Social Care Trust in Northern Ireland}
  \item{Able to participate in the NTS?}{Is this LA/HSCT participating in the National Transfer Scheme?}
  \item{Transfers out of Local Authority}{Number of transfers out of LA}
  \item{Transfers into Local Authority}{Number of transfers into LA}
  \item{Transfers into to LA from port/intake unit}{Number of transfers into to LA from port/intake unit}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
nts
}
\description{
Number of individuals transferred between local authorities under the
National Transfer Scheme for unaccompanied asylum seeking children.
}
\keyword{datasets}
