% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oral_hearings_volume}
\alias{oral_hearings_volume}
\title{Volume of oral hearings represented by the Home Office}
\format{
A data frame with 12 rows and 4 variables:
\describe{
  \item{Financial Year}{Financial year}
  \item{Asylum appeals represented at hearing (\%)}{Percent of sylum appeals represented at hearing}
  \item{Asylum Oral Hearings}{Number of asylum oral hearings}
  \item{Asylum Oral Hearings Represented by HO}{Number of asylum oral hearings represented by the Home Office}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
oral_hearings_volume
}
\description{
The volume of oral hearings represented by the Home Office divided by the
total number of oral hearing.
}
\keyword{datasets}
