% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outcomes}
\alias{outcomes}
\title{Outcome analysis of asylum applications}
\format{
A data frame with 26 variables:
\describe{
  \item{Year of Application}{Year of application}
  \item{Geographical Region}{Geographical region}
  \item{Nationality}{Nationality}
  \item{Applications}{Applications}
  \item{Initial Decisions}{Initial decisions}
  \item{Initial: Grants of Protection}{Grants of protection at initial decision}
  \item{Initial: Grants of Other Leave}{Grants of other leave at initial decision}
  \item{Initial: Refusals}{Refusals at initial decision}
  \item{Initial: Withdrawn Applications}{Withdrawn applications at initial decision}
  \item{Initial: Administrative Outcomes}{Administrative outcomes at initial decision}
  \item{Initial: Not Yet Known}{Not yet known}
  \item{Enforced Returns}{Enforced returns}
  \item{Voluntary Returns}{Voluntary returns}
  \item{Assisted Returns (Subset of Voluntary returns)}{Assisted Returns (Subset of Voluntary returns)}
  \item{Controlled Returns (Subset of Voluntary returns)}{Controlled Returns (Subset of Voluntary returns)}
  \item{Other Verified Returns (Subset of Voluntary returns)}{Other Verified Returns (Subset of Voluntary returns)}
  \item{Latest: Grants of Protection}{Latest: Grants of Protection}
  \item{Latest: Grants of Other Leave}{Latest: Grants of Other Leave}
  \item{Latest: Refusals}{Latest: Refusals}
  \item{Latest: Withdrawn Applications}{Latest: Withdrawn Applications}
  \item{Latest: Administrative Outcomes}{Latest: Administrative Outcomes}
  \item{Latest: Not Yet Known}{Latest: Not Yet Known}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/asylum-and-resettlement-datasets}
}
\usage{
outcomes
}
\description{
The initial decision and final outcome of all asylum applications raised in
a period, by nationality.
}
\keyword{datasets}
