% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_entering_detention}
\alias{people_entering_detention}
\title{Number of entries into immigration detention}
\format{
A data frame with 70,342 rows and 9 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality}
  \item{Region}{Region}
  \item{Sex}{Sex}
  \item{Age}{Age}
  \item{First place of detention}{First place of detention}
  \item{Entering}{Entering}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/returns-and-detention-datasets}
}
\usage{
people_entering_detention
}
\description{
Number of entries into immigration detention by nationality, age, sex and
initial place of detention.
}
\keyword{datasets}
