% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_in_detention}
\alias{people_in_detention}
\title{Number of people in immigration detention}
\format{
A data frame with 77,146 rows and 10 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality}
  \item{Region}{Region}
  \item{Sex}{Sex}
  \item{Age}{Age}
  \item{Current place of detention}{Current place of detention}
  \item{Length of detention}{Length of detention}
  \item{People}{Number of people in detention}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/returns-and-detention-datasets}
}
\usage{
people_in_detention
}
\description{
Number of people in immigration detention at the end of each quarter by
nationality, age, sex, current place of detention and length of detention.
}
\keyword{datasets}
