% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_leaving_detention}
\alias{people_leaving_detention}
\title{Number of occurrences of people leaving detention}
\format{
A data frame with 100,886 rows and 10 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality}
  \item{Region}{Region}
  \item{Sex}{Sex}
  \item{Age}{Age}
  \item{Reason for leaving detention}{Reason for leaving detention}
  \item{Length of detention}{Length of detention}
  \item{Leaving}{Number of people leaving detention}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/returns-and-detention-datasets}
}
\usage{
people_leaving_detention
}
\description{
Number of occurrences of people leaving detention by nationality, age, sex,
reason for leaving detention and length of detention
}
\keyword{datasets}
