% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns}
\alias{returns}
\title{Number of returns from the UK}
\format{
A data frame with 227,607 rows and 11 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality}
  \item{Region}{Region}
  \item{Return type group}{Return type group}
  \item{Return type}{Return type}
  \item{Age}{Age}
  \item{Sex}{Sex}
  \item{Return destination group}{Return destination group}
  \item{Number of returns}{Number of returns}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/returns-and-detention-datasets}
}
\usage{
returns
}
\description{
Number of returns from the UK, by nationality, age, sex, type of return and
return destination group
}
\keyword{datasets}
