% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_asylum_longitudinal}
\alias{returns_asylum_longitudinal}
\title{Returns from the UK, by nationality, type of return, and asylum/non-asylum}
\format{
A data frame with 26 rows and 5 variables:
\describe{
  \item{Date of return}{Date of return}
  \item{Category}{Asylum-related or non-asylum-related return?}
  \item{Enforced returns}{Number of enforced returns}
  \item{Voluntary returns}{Number of voluntary returns}
  \item{Refused entry at port and subsequently departed}{Number refused entry at port and subsequently departed}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/returns-and-detention-datasets}
}
\usage{
returns_asylum_longitudinal
}
\description{
Returns from the UK, by nationality, type of return, and asylum/non-asylum
}
\keyword{datasets}
