% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_by_destination}
\alias{returns_by_destination}
\title{Number of returns from the UK, by destination}
\format{
A data frame with 41,500 rows and 8 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Return type group}{Return type group}
  \item{Return type}{Return type}
  \item{Return destination}{Return destination}
  \item{Return destination region}{Return destination region}
  \item{Number of returns}{Number of returns}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/returns-and-detention-datasets}
}
\usage{
returns_by_destination
}
\description{
Number of returns from the UK, by type of return and country of destination
}
\keyword{datasets}
