% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_offenders_by_destination}
\alias{returns_offenders_by_destination}
\title{Number of foreign national offender returns from the UK, by destination}
\format{
A data frame with 4,412 rows and 6 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Return destination}{Return destination}
  \item{Return destination region}{Return destination region}
  \item{Number of foreign national offender returns}{Number of foreign national offender returns}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/returns-and-detention-datasets}
}
\usage{
returns_offenders_by_destination
}
\description{
Number of foreign national offender returns from the UK, by destination.
}
\keyword{datasets}
