% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_offenders_by_nationality}
\alias{returns_offenders_by_nationality}
\title{Number of foreign national offender returns from the UK, by nationality}
\format{
A data frame with 6,217 rows and 7 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality}
  \item{Region}{Region}
  \item{Return destination group}{Return destination group}
  \item{Number of foreign national offender returns}{Number of foreign national offender returns}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/returns-and-detention-datasets}
}
\usage{
returns_offenders_by_nationality
}
\description{
Number of foreign national offender returns from the UK, by nationality and
return destination group.
}
\keyword{datasets}
