% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-db.R
\name{save_sqlite}
\alias{save_sqlite}
\title{Save all the asylum and resettlement data in a SQLite database}
\usage{
save_sqlite(db_path, db_filename = "asylum.sqlite")
}
\arguments{
\item{db_path}{Where to save the SQLite database}

\item{db_filename}{SQLite filename}
}
\description{
Save all the asylum and resettlement data in a SQLite database
}
