% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_boat_asylum_applications}
\alias{small_boat_asylum_applications}
\title{Number of asylum applications from small boat arrivals}
\format{
A data frame with 11 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Asylum application}{Was an asylum application raised?}
  \item{Nationality}{Nationality}
  \item{Region}{Region}
  \item{Applicant type}{Main applicant or dependant}
  \item{UASC}{Unaccompanied asylum-seeking children?}
  \item{Sex}{Sex}
  \item{Age Group}{Age Group}
  \item{Applications}{Number of applications}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/irregular-migration-detailed-dataset-and-summary-tables}
}
\usage{
small_boat_asylum_applications
}
\description{
Number of asylum applications from small boat arrivals by asylum application,
nationality, region, applicant type, sex and age group.
}
\keyword{datasets}
