% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_boat_initial_decisions}
\alias{small_boat_initial_decisions}
\title{Initial decisions on asylum applications from small boat arrivals}
\format{
A data frame with 4,365 rows and 11 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality}
  \item{Region}{Region}
  \item{Asylum Case Outcome}{Specific outcome of the asylum case}
  \item{Asylum Case Outcome Group}{Type of outcome}
  \item{Applicant type}{Main applicant or dependant}
  \item{Sex}{Sex}
  \item{Age Group}{Age Group}
  \item{Decisions}{Number of decisions}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/irregular-migration-detailed-dataset-and-summary-tables}
}
\usage{
small_boat_initial_decisions
}
\description{
Initial decisions on asylum applications from small boat arrivals by
nationality, region, asylum case outcome, asylum case outcome group,
applicant type, sex and age group.
}
\keyword{datasets}
