% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_boat_NRM_outcomes}
\alias{small_boat_NRM_outcomes}
\title{Outcomes of NRM referrals from small boat arrivals}
\format{
A data frame with 2,768 rows and 10 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Nationality}{Nationality}
  \item{Region}{Region}
  \item{Sex}{Sex}
  \item{Age Group}{Age Group}
  \item{NRM Case Outcome}{Specific outcome of the NRM case}
  \item{NRM Case Outcome Group}{Type of outcome}
  \item{Referrals}{Number of referrals}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/irregular-migration-detailed-dataset-and-summary-tables}
}
\usage{
small_boat_NRM_outcomes
}
\description{
Outcomes of NRM referrals from small boat arrivals by nationality, region,
sex, age group, NRM case outcome and NRM case outcome group.
}
\keyword{datasets}
