% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_boat_NRM_referrals}
\alias{small_boat_NRM_referrals}
\title{Number of NRM referrals from small boat arrivals}
\format{
A data frame with 9 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{NRM referral}{Was an NRM referral raised?}
  \item{Nationality}{Nationality}
  \item{Region}{Region}
  \item{Sex}{Sex}
  \item{Age Group}{Age Group}
  \item{Referrals}{Number of referrals}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/irregular-migration-detailed-dataset-and-summary-tables}
}
\usage{
small_boat_NRM_referrals
}
\description{
Number of NRM referrals from small boat arrivals by NRM referral,
nationality, region, sex and age group.
}
\keyword{datasets}
