% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{support_applications}
\alias{support_applications}
\title{Applications for section 95 support}
\format{
A data frame with 4,324 rows and 6 variables:
\describe{
  \item{Year}{Year}
  \item{Nationality}{Nationality of people receiving support}
  \item{Region}{World region the nation is in}
  \item{Support type granted}{'Accommodation', 'Closed / Invalid', or 'Subsistence Only'}
  \item{Group type}{'Single adult', 'Family', or 'Unknown'}
  \item{Applications}{Number of applications}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/asylum-and-resettlement-datasets}
}
\usage{
support_applications
}
\description{
Applications for section 95 support, by nationality, support type granted,
and group type.
}
\keyword{datasets}
