% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{support_applications_rasi}
\alias{support_applications_rasi}
\title{Applications for Support Section 4 and Section 95}
\format{
A data frame with 50 rows and 5 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{Section 4}{Number of applications for Section 4 support}
  \item{Section 95}{Number of applications for Section 95 support}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
support_applications_rasi
}
\description{
Applications for Support Section 4 and Section 95.
}
\keyword{datasets}
