% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{travel_documents}
\alias{travel_documents}
\title{Travel documents}
\format{
A data frame with 49 rows and 5 variables:
\describe{
  \item{Date}{Date}
  \item{Year}{Year}
  \item{Quarter}{Quarter}
  \item{TD Raised}{Number of travel document applications raised}
  \item{TD Despatched}{Number of travel documents despatched}
  ...
}
}
\source{
\url{https://www.gov.uk/government/collections/migration-transparency-data#uk-visas-and-immigration}
}
\usage{
travel_documents
}
\description{
Travel documents, applications raised and despatched.
}
\keyword{datasets}
