% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{children_entering_detention}
\alias{children_entering_detention}
\title{Children entering detention}
\format{
A data frame with 13 rows and 7 variables:
\describe{
  \item{Date of entry to detention}{Date of entry to detention}
  \item{Total children}{Total children}
  \item{Children Under 5 yrs}{Children Under 5 yrs}
  \item{Children 5-11 yrs}{Children 5-11 yrs}
  \item{Children 12-16 yrs}{Children 12-16 yrs}
  \item{Children 17 yrs}{Children 17 yrs}
  \item{Total adults (18+)}{Total adults (18+)}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/immigration-system-statistics-data-tables-year-ending-december-2022#detention}
}
\usage{
children_entering_detention
}
\description{
Number of occurrences of children entering detention, 2010 to 2022
}
\keyword{datasets}
