% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detentions_deaths_and_absconds}
\alias{detentions_deaths_and_absconds}
\title{Deaths of persons detained under Immigration Act Powers and absconds from
detention, 2017 to 2021}
\format{
A data frame with 5 rows and 4 variables:
\describe{
  \item{Year}{Year}
  \item{Deaths of persons detained under Immigration Powers}{Deaths of persons detained under Immigration Powers}
  \item{Of which occurred in the detention estate}{Of which occurred in the detention estate}
  \item{Absconds from detention}{Absconds from detention}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/immigration-system-statistics-data-tables-year-ending-december-2022#detention}
}
\usage{
detentions_deaths_and_absconds
}
\description{
Deaths of persons detained under Immigration Act Powers and absconds from
detention, 2017 to 2021
}
\keyword{datasets}
