% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outcomes}
\alias{outcomes}
\title{Outcome analysis of asylum applications}
\format{
A data frame with 3,463 rows and 26 variables:
\describe{
  \item{Year of application}{Year of application}
  \item{Geographical region}{Geographical region}
  \item{Nationality}{Nationality}
  \item{Applications}{Applications}
  \item{Initial decisions}{Initial decisions}
  \item{Granted asylum}{Granted asylum}
  \item{Granted HP/DL}{Granted HP/DL}
  \item{Other grants}{Other grants}
  \item{Refused}{Refused}
  \item{Applications withdrawn}{Applications withdrawn}
  \item{Initial decisions not yet known}{Initial decisions not yet known}
  \item{Appeals lodged}{Appeals lodged}
  \item{Allowed appeals}{Allowed appeals}
  \item{Dismissed appeals}{Dismissed appeals}
  \item{Withdrawn appeals}{Withdrawn appeals}
  \item{Appeal outcome not yet known}{Appeal outcome not yet known}
  \item{Enforced returns}{Enforced returns}
  \item{Voluntary returns}{Voluntary returns}
  \item{Assisted returns (Subset of Voluntary returns)}{Assisted returns (Subset of Voluntary returns)}
  \item{Controlled returns (Subset of Voluntary returns)}{Controlled returns (Subset of Voluntary returns)}
  \item{Other verified returns (Subset of Voluntary returns)}{Other verified returns (Subset of Voluntary returns)}
  \item{Final outcome: Grants of asylum}{Final outcome: Grants of asylum}
  \item{Final outcomes: Grants of HP/DL and other}{Final outcomes: Grants of HP/DL and other}
  \item{Final outcomes: Refused asylum or HP or DL or other leave}{Final outcomes: Refused asylum or HP or DL or other leave}
  \item{Final outcome: Withdrawn application}{Final outcome: Withdrawn application}
  \item{Final outcome: Cases with final outcome not yet known}{Final outcome: Cases with final outcome not yet known}
  ...
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/asylum-and-resettlement-datasets}
}
\usage{
outcomes
}
\description{
The initial decision and final outcome of all asylum applications raised in
a period, by nationality.
}
\keyword{datasets}
